refactor export button
